+++ data/src/main/java/net/pantasystem/milktea/data/di/module/DriveDirectoryModule.kt
@@ -1,12 +1,12 @@
-package jp.panta.misskeyandroidclient.di.module
+package net.pantasystem.milktea.data.di.module
 
 import dagger.Binds
 import dagger.Module
 import dagger.hilt.InstallIn
 import dagger.hilt.components.SingletonComponent
+import net.pantasystem.milktea.app_store.drive.DriveDirectoryPagingStore
 import net.pantasystem.milktea.data.infrastructure.drive.DriveDirectoryPagingStoreImpl
 import net.pantasystem.milktea.data.infrastructure.drive.DriveDirectoryRepositoryImpl
-import net.pantasystem.milktea.app_store.drive.DriveDirectoryPagingStore
 import net.pantasystem.milktea.model.drive.DriveDirectoryRepository
 import javax.inject.Singleton
