+++ data/src/main/java/net/pantasystem/milktea/data/di/module/MetaModule.kt
@@ -1,19 +1,19 @@
-package jp.panta.misskeyandroidclient.di.module
+package net.pantasystem.milktea.data.di.module
 
 import dagger.Module
 import dagger.Provides
 import dagger.hilt.InstallIn
 import dagger.hilt.components.SingletonComponent
+import net.pantasystem.milktea.common.Logger
 import net.pantasystem.milktea.data.api.misskey.MisskeyAPIProvider
 import net.pantasystem.milktea.data.infrastructure.DataBase
-import net.pantasystem.milktea.model.instance.MediatorFetchMeta
-import net.pantasystem.milktea.model.instance.MetaRepository
-import net.pantasystem.milktea.model.instance.FetchMeta
 import net.pantasystem.milktea.data.infrastructure.instance.db.InMemoryMetaRepository
 import net.pantasystem.milktea.data.infrastructure.instance.db.MediatorMetaRepository
 import net.pantasystem.milktea.data.infrastructure.instance.db.RoomMetaRepository
 import net.pantasystem.milktea.data.infrastructure.instance.remote.RemoteFetchMeta
-import net.pantasystem.milktea.common.Logger
+import net.pantasystem.milktea.model.instance.FetchMeta
+import net.pantasystem.milktea.model.instance.MediatorFetchMeta
+import net.pantasystem.milktea.model.instance.MetaRepository
 import javax.inject.Singleton
 
 @Module
